chore(mobile): legacy quarantine wave 2 — lib/src + main.dart + old l10n into lib/legacy (§11)
git-mv'd structure-preserved, never deleted (the web quarantine pattern):
all of lib/src/** (auth surface, legacy theme trio, localization core,
persistence, shared widgets), the superseded main.dart, and the old l10n
files (app_sq.arb + the committed generated localizations) now live under
mobile/flutter/lib/legacy/ — outside analysis, codegen, CI, and builds.

The one KEEP-register salvage rides along: countdown.dart moves live to
src/core/ui/ (C6's 3-2-1 countdown), touched only to clear the analyzer
gates — behavior unchanged.
