+++ mobile/flutter/lib/legacy/src/app/splash_screen.dart
