+++ mobile/flutter/lib/legacy/src/app/welcome_screen.dart
