+++ mobile/flutter/lib/legacy/src/features/auth/sign_up_form.dart
