+++ mobile/flutter/lib/legacy/src/features/auth/sign_up_screen.dart
