+++ mobile/flutter/lib/legacy/src/features/auth/verify_email.dart
