+++ mobile/flutter/lib/legacy/src/shared/models/user.dart
