+++ mobile/flutter/lib/legacy/src/shared/number_text_input_formatter.dart
